исправление замечаний домашка 12
