+++ src/main/java/ru/learningJava/homework12/Sex.java
@@ -1,6 +1,6 @@
 package ru.learningJava.homework12;
 
 public enum Sex {
-    m,
-    f
+    Male,
+    Female
 }
