+++ src/main/java/ru/learningJava/homework12/securityValidation/User.java
@@ -24,13 +24,15 @@ public class User {
     }
 
     public static void checkLogin(String login){
-        if (!login.matches("^([A-Za-z0-9_]{0,20})$")) {
+        String regex = "^([A-Za-z0-9_]{0,20})$";
+        if (!login.matches(regex)) {
             throw new WrongLoginException("Логин не соответствует формату");
         }
     }
 
     public static void checkPassword(String password, String confirmPassword){
-        if (!password.matches("^([A-Za-z0-9_]{0,20})$")) {
+        String regex = "^([A-Za-z0-9_]{0,20})$";
+        if (!password.matches(regex)) {
             throw new WrongPasswordException("Пароль не соответствует формату");
         }
